+++ changelog.md
@@ -5,9 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.1.7] - 2018-
+
+### Added
+- Use TkAgg backend in matplotlib for consistent performance
+- Download 2017 logs in [kgs_download](https://toddheitmann.github.io/PetroPy/function/download.html#petropy.kgs_download) function
+
+### Fixed
+- [Issue 1](https://github.com/toddheitmann/PetroPy/issues/1) calling newer methods from old versions of required packages. Versions added to requirements.txt.
+
 ## [0.1.6] - 2018-02-15
 
-## Fixed
+### Fixed
 - Issue with show method for LogViewer class using different back ends of matplotlib
 
 ## [0.1.5] - 2017-10-24